v1/functions/user_roles/requirements.txt
orator
PyMySQL
PyJWT
